+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - Errors occurring while initially loading databases
       (or when calling `Geolix.reload_databases/0`) are now sent to
       `Logger.error` ([#16](https://github.com/elixir-geolix/geolix/issues/16))
+    - When an error occurs during database load the database will be
+      excluded from the lookup process
 
 ## v0.13.0 (2017-04-12)