+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.14.0-dev
 
 - Enhancements
+    - Configuration errors are logged and/or returned
     - Databases in tarball format (`.tar` or `.tar.gz`) are now supported
       ([#16](https://github.com/elixir-geolix/geolix/issues/16))
     - Errors are now always returned in a tuple format (`{ :error, type }`)